backend/ToDoRNA/requirements.txt
Django==2.0.3
djangorestframework==3.7.7
jsonfield==2.0.2
pytz==2018.3
